Extract out HTTP related `framework` types into the new `http` component (athena-framework/athena#640)
## Context

Should have likely did this long ago as it also handles the code smell
of using the `contracts` component as a place to also store "common"
things when they really should have just all been in an `http`
component. The `framework` component at this point is more similar to a
`framework-bundle`. TBD if we actually want to rename the shard itself
to make that more clear at some point. ATM I don't think there'd be much
benefit in doing so.

This is a fairly wide-spread change as it involves renaming very common
types, but should be fairly easy to migrate via some find/replaces.

Resolves https://github.com/athena-framework/athena/issues/627

Had to add `::` to `HTTP` types so they resolved correctly and not to
`Athena::HTTP`.

## Changelog

- Extract out HTTP related `framework` types into the new `http`
component